put roc to probabilistic
