+++ xskillscore/tests/test_probabilistic.py
@@ -5,6 +5,7 @@ import pytest
 import xarray as xr
 from scipy.stats import norm
 from sklearn.calibration import calibration_curve
+from sklearn.metrics import roc_auc_score, roc_curve
 from xarray.tests import assert_allclose, assert_identical
 
 from xskillscore.core.probabilistic import (
@@ -15,6 +16,7 @@ from xskillscore.core.probabilistic import (
     discrimination,
     rank_histogram,
     reliability,
+    roc,
     rps,
     threshold_brier_score,
 )
@@ -474,3 +476,137 @@ def test_rps_perfect_values(o, category_edges):
 def test_rps_dask(o_dask, f_prob_dask, category_edges):
     """Test that rps returns dask array if provided dask array"""
     assert rps(o_dask, f_prob_dask, category_edges=category_edges).chunks is not None
+
+
+@pytest.mark.parametrize(
+    "forecast, observation",
+    [
+        (
+            pytest.lazy_fixture("forecast_1d_long"),
+            pytest.lazy_fixture("observation_1d_long"),
+        ),
+        (
+            pytest.lazy_fixture("forecast_3d"),
+            pytest.lazy_fixture("observation_3d"),
+        ),
+    ],
+)
+@pytest.mark.parametrize("dim", ["time", None])
+@pytest.mark.parametrize("drop_intermediate", [True, False])
+@pytest.mark.parametrize("chunk", [True, False])
+@pytest.mark.parametrize("input", ["Dataset", "DataArray"])
+@pytest.mark.parametrize(
+    "return_results", ["all_as_tuple", "area", "all_as_metric_dim"]
+)
+def test_roc_returns(
+    forecast,
+    observation,
+    symmetric_edges,
+    dim,
+    return_results,
+    input,
+    chunk,
+    drop_intermediate,
+):
+    """testing keywords and inputs"""
+    if "Dataset" in input:
+        name = "var"
+        forecast = forecast.to_dataset(name=name)
+        observation = observation.to_dataset(name=name)
+    if chunk:
+        forecast = forecast.chunk()
+        observation = observation.chunk()
+
+    roc(
+        forecast,
+        observation,
+        symmetric_edges,
+        dim=dim,
+        drop_intermediate=drop_intermediate,
+        return_results=return_results,
+    )
+
+
+def test_roc_auc_score_random_forecast(
+    forecast_1d_long, observation_1d_long, symmetric_edges
+):
+    """Test that ROC AUC around 0.5 for random forecast."""
+    area = roc(
+        forecast_1d_long,
+        observation_1d_long,
+        symmetric_edges,
+        dim="time",
+        return_results="area",
+    )
+    assert area < 0.6
+    assert area > 0.4
+
+
+def test_roc_auc_score_perfect_forecast(forecast_1d_long, symmetric_edges):
+    """Test that ROC AUC equals 1 for perfect forecast."""
+    area = roc(
+        forecast_1d_long,
+        forecast_1d_long,
+        symmetric_edges,
+        dim="time",
+        return_results="area",
+    )
+    assert area == 1.0
+
+
+def test_roc_auc_score_constant_forecast(forecast_1d_long, symmetric_edges):
+    """Test that ROC AUC equals 0 or 0.5. for constant forecast."""
+    area = roc(
+        forecast_1d_long,
+        xr.ones_like(forecast_1d_long) * 10,
+        symmetric_edges,
+        dim="time",
+        return_results="area",
+    )
+    assert float(area) in [0.0, 0.5]
+
+
+@pytest.mark.parametrize("drop_intermediate", [False, True])
+def test_roc_bin_edges_continuous_against_sklearn(
+    forecast_1d_long, observation_1d_long, drop_intermediate
+):
+    """Test xs.roc against sklearn.metrics.roc_curve/auc_score."""
+    fb = forecast_1d_long > 0  # binary
+    op = np.clip(observation_1d_long, 0, 1)  # prob
+    # sklearn
+    sk_fpr, sk_tpr, _ = roc_curve(fb, op, drop_intermediate=drop_intermediate)
+    sk_area = roc_auc_score(fb, op)
+    # xs
+    xs_fpr, xs_tpr, xs_area = roc(
+        fb,
+        op,
+        "continuous",
+        drop_intermediate=drop_intermediate,
+        return_results="all_as_tuple",
+    )
+    np.testing.assert_allclose(xs_area, sk_area)
+    if not drop_intermediate:  # drops sometimes one too much or too little
+        assert (xs_fpr == sk_fpr).all()
+        assert (xs_tpr == sk_tpr).all()
+
+
+def test_roc_bin_edges_drop_intermediate(forecast_1d_long, observation_1d_long):
+    """Test that drop_intermediate reduces probability_bins in xs.roc ."""
+    fb = forecast_1d_long > 0  # binary
+    op = np.clip(observation_1d_long, 0, 1)  # prob
+    # xs
+    txs_fpr, txs_tpr, txs_area = roc(
+        fb, op, "continuous", drop_intermediate=True, return_results="all_as_tuple"
+    )
+    fxs_fpr, fxs_tpr, fxs_area = roc(
+        fb, op, "continuous", drop_intermediate=False, return_results="all_as_tuple"
+    )
+    # same area
+    np.testing.assert_allclose(fxs_area, txs_area)
+    # same or less probability_bins
+    assert len(fxs_fpr) >= len(txs_fpr)
+    assert len(fxs_tpr) >= len(txs_tpr)
+
+
+def test_roc_multi_dim(forecast_3d, observation_3d):
+    roc(forecast_3d, observation_3d, bin_edges=np.linspace(0, 10 + 1e-8, 6))
